Refactor ChatInterface and Index files
Refactor logic from `src/client/ChatInterface.tsx` and `src/client/Index.tsx` into smaller files for better readability and management.

* **Real-time Communication Logic**:
  - Move real-time communication logic to `src/client/RealTimeComm.tsx`.
  - Implement `usePartySocket` function to handle messages and authentication.

* **User Authentication Logic**:
  - Move user authentication logic to `src/client/UserAuth.tsx`.
  - Implement `handleAuthCallback` function to handle authentication callback and set user profile.

* **UI Components**:
  - Create `src/client/UIComponents.tsx` and move UI components like `FeedbackButtons`, `RatingSystem`, `FeedbackForm`, `SurveyLink`, `EmbeddedSurvey`, `MessageReactions`, `MessageThreads`, `AnalyticsDisplay`, and `CloudflareSecretsSettings` to this file.

* **Routing and Error Handling**:
  - Create `src/client/Routing.tsx` for routing logic.
  - Create `src/client/ErrorHandling.tsx` for error handling logic.

* **Update `src/client/ChatInterface.tsx`**:
  - Remove user authentication logic and import `handleAuthCallback` from `src/client/UserAuth.tsx`.
  - Remove real-time communication logic and import `usePartySocket` from `src/client/RealTimeComm.tsx`.
  - Remove UI components and import them from `src/client/UIComponents.tsx`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Derpenstein69/derp-chat?shareId=XXXX-XXXX-XXXX-XXXX).
