+++ src/client/components/AnalyticsDisplay.tsx
@@ -1,0 +1,72 @@
+import React, { useState, useEffect } from "react";
+
+export function AnalyticsDisplay(): JSX.Element {
+  const [analyticsData, setAnalyticsData] = useState<{
+    messageCount: number;
+    userActivity: { [key: string]: number };
+    messageFrequency: number;
+    userPreferences: { [key: string]: any };
+    previousInteractions: { [key: string]: any };
+    activeSessions: number;
+    sessionDuration: number;
+  }>({
+    messageCount: 0,
+    userActivity: {},
+    messageFrequency: 0,
+    userPreferences: {},
+    previousInteractions: {},
+    activeSessions: 0,
+    sessionDuration: 0,
+  });
+
+  useEffect(() => {
+    const handleAnalyticsUpdate = (evt: MessageEvent) => {
+      const data = JSON.parse(evt.data);
+      setAnalyticsData(data);
+    };
+
+    const analyticsSocket = new WebSocket("wss://your-analytics-endpoint");
+    analyticsSocket.addEventListener("message", handleAnalyticsUpdate);
+
+    return () => {
+      analyticsSocket.removeEventListener("message", handleAnalyticsUpdate);
+      analyticsSocket.close();
+    };
+  }, []);
+
+  return (
+    <div className="analytics-display">
+      <h3>Real-Time Analytics</h3>
+      <p>Message Count: {analyticsData.messageCount}</p>
+      <p>Message Frequency: {analyticsData.messageFrequency} messages/minute</p>
+      <h4>User Activity</h4>
+      <ul>
+        {Object.entries(analyticsData.userActivity).map(([user, activity]) => (
+          <li key={user}>
+            {user}: {activity} messages
+          </li>
+        ))}
+      </ul>
+      <h4>User Preferences</h4>
+      <ul>
+        {Object.entries(analyticsData.userPreferences).map(([user, preferences]) => (
+          <li key={user}>
+            {user}: {JSON.stringify(preferences)}
+          </li>
+        ))}
+      </ul>
+      <h4>Previous Interactions</h4>
+      <ul>
+        {Object.entries(analyticsData.previousInteractions).map(([user, interactions]) => (
+          <li key={user}>
+            {user}: {JSON.stringify(interactions)}
+          </li>
+        ))}
+      </ul>
+      <h4>Active Sessions</h4>
+      <p>Active Sessions: {analyticsData.activeSessions}</p>
+      <h4>Session Duration</h4>
+      <p>Session Duration: {analyticsData.sessionDuration} minutes</p>
+    </div>
+  );
+}
