+++ src/client/components/FeedbackButtons.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export function FeedbackButtons({ messageId }: { messageId: string }): JSX.Element {
+  const handleFeedback = (feedback: string) => {
+    console.log(`Feedback for message ${messageId}: ${feedback}`);
+  };
+
+  return (
+    <div className="feedback-buttons">
+      <button onClick={() => handleFeedback("thumbs-up")}>👍</button>
+      <button onClick={() => handleFeedback("thumbs-down")}>👎</button>
+    </div>
+  );
+}
