+++ src/client/components/SurveyLink.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export function SurveyLink(): JSX.Element {
+  const handleSurveyClick = () => {
+    console.log("Survey link clicked");
+  };
+
+  return (
+    <a className="survey-link" onClick={handleSurveyClick}>
+      Take our survey
+    </a>
+  );
+}
